pyproject.toml
[tool.poetry]
name = "lobsterobot"
version = "1.0.0"
description = "A bad discord bot. Written by Kaylynn234/ Kaylynn#4444."
authors = ["Kaylynn <mkaylynn7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord-ext-menus" = { git = "https://github.com/Rapptz/discord-ext-menus.git", branch = "master" }
chattymarkov = "^1.2.7"
matplotlib = "^3.1.2"
spotipy = "^2.7.1"
nltk = "^3.4.5"
dateparser = "^0.7.2"
demoji = "^0.1.5"
pendulum = "^2.0.5"
aiohttp = "^3.6.2"
regex = "^2020.1.8"
dataset = "^1.2.0"
pillow = "^7.0.0"
emoji = "^0.5.4"
jishaku = "^1.17.1"
"discord.py" = {extras = ["voice"], version = "^1.3.1"}
wkhtmltopdf = "^0.2"
imgkit = "^1.0.2"
validators = "^0.14.2"
natural = "^0.2.0"
py_expression_eval = "^0.3.9"
requests = "^2.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
